refactored pipeline to be linear.
so this change does make any changes to the tasks themselves, it only
updates a couple triggers. however, it is a major overhaul on process.

now, in order to progress through the pipeline, every job must complete
successfully. this aligns with master-in-production deployment style
that lets us make very small, very quick changes, and when it reaches
production, it's been vetted by 3 bosh deployments and 9 different test
suites.

in order to ensure the pipeline is human-readable with a low cognitive
overhead, every job's connection is with `pipeline-tasks`.
`pipeline-tasks` is not used by every job, but it forces concourse's d3
configuration to render the pipeline in a very clean, clear, concise
method. it also allows for an "aggregation step", so each downstream job
won't start until all dependencies of `pipeline-tasks` have completed
successfully. the addition and/or new dependency on `pipeline-tasks` is
purely so the pipeline is human-friendly and has no impact on the job
itself.

Signed-off-by: Mike Lloyd <mike.lloyd@gsa.gov>